refactor do package data para network da apod
