+++ app/src/main/java/com/nasinha/digitalspace/apod/repository/ApodEndPoint.kt
@@ -1,6 +1,6 @@
 package com.nasinha.digitalspace.apod.repository
 
-import com.nasinha.digitalspace.apod.data.NetworkUtilsApod
+import com.nasinha.digitalspace.apod.network.NetworkUtilsApod
 import com.nasinha.digitalspace.apod.model.ApodResponseModel
 import retrofit2.http.GET
 import retrofit2.http.Query
